+++ ijkmedia/ijksdl/gles2/fsh/yuv420p.fsh.c
@@ -1,9 +1,5 @@
 /*
- * IJKSDLGLShader.m
- *
- * Copyright (c) 2013 Zhang Rui <bbcallen@gmail.com>
- *
- * based on https://github.com/kolyvan/kxmovie
+ * copyright (c) 2016 Zhang Rui <bbcallen@gmail.com>
  *
  * This file is part of ijkPlayer.
  *
@@ -22,19 +18,30 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#import "IJKSDLGLShader.h"
+#include "ijksdl/gles2/internal.h"
 
-//////////////////////////////////////////////////////////
-
-#pragma mark - shaders
-
-NSString *const rgbFragmentShaderString = IJK_SHADER_STRING
-(
-    varying highp vec2 v_texcoord;
-    uniform sampler2D s_texture;
+static const char g_shader[] = IJK_GLES_STRING(
+    precision highp float;
+    varying   highp vec2 vv2_Texcoord;
+    uniform         mat3 um3_ColorConversion;
+    uniform   lowp  sampler2D us2_SamplerX;
+    uniform   lowp  sampler2D us2_SamplerY;
+    uniform   lowp  sampler2D us2_SamplerZ;
 
     void main()
     {
-        gl_FragColor = texture2D(s_texture, v_texcoord);
+        mediump vec3 yuv;
+        lowp    vec3 rgb;
+
+        yuv.x = (texture2D(us2_SamplerX, vv2_Texcoord).r - (16.0 / 255.0));
+        yuv.y = (texture2D(us2_SamplerY, vv2_Texcoord).r - 0.5);
+        yuv.z = (texture2D(us2_SamplerZ, vv2_Texcoord).r - 0.5);
+        rgb = um3_ColorConversion * yuv;
+        gl_FragColor = vec4(rgb, 1);
     }
 );
+
+const char *IJK_GLES2_getFragmentShader_yuv420p()
+{
+    return g_shader;
+}