+++ ijkmedia/ijksdl/gles2/fsh/yuv420sp.fsh.c
@@ -1,9 +1,5 @@
 /*
- * IJKSDLGLShader.h
- *
- * Copyright (c) 2013 Zhang Rui <bbcallen@gmail.com>
- *
- * based on https://github.com/kolyvan/kxmovie
+ * copyright (c) 2016 Zhang Rui <bbcallen@gmail.com>
  *
  * This file is part of ijkPlayer.
  *
@@ -22,8 +18,28 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#import <Foundation/Foundation.h>
+#include "ijksdl/gles2/internal.h"
 
-#define IJK_STRINGIZE(x) #x
-#define IJK_STRINGIZE2(x) IJK_STRINGIZE(x)
-#define IJK_SHADER_STRING(text) @ IJK_STRINGIZE2(text)
+static const char g_shader[] = IJK_GLES_STRING(
+    precision highp float;
+    varying   highp vec2 vv2_Texcoord;
+    uniform         mat3 um3_ColorConversion;
+    uniform   lowp  sampler2D us2_SamplerX;
+    uniform   lowp  sampler2D us2_SamplerY;
+
+    void main()
+    {
+        mediump vec3 yuv;
+        lowp    vec3 rgb;
+
+        yuv.x  = (texture2D(us2_SamplerX,  vv2_Texcoord).r  - (16.0 / 255.0));
+        yuv.yz = (texture2D(us2_SamplerY,  vv2_Texcoord).rg - vec2(0.5, 0.5));
+        rgb = um3_ColorConversion * yuv;
+        gl_FragColor = vec4(rgb, 1);
+    }
+);
+
+const char *IJK_GLES2_getFragmentShader_yuv420sp()
+{
+    return g_shader;
+}