+++ ijkmedia/ijksdl/ijksdl_fourcc.h
@@ -55,6 +55,8 @@
 #define SDL_FCC_UYVY    SDL_FOURCC('U', 'Y', 'V', 'Y')  /**< bpp=16, Packed mode: U0+Y0+V0+Y1 (1 plane) */
 #define SDL_FCC_YVYU    SDL_FOURCC('Y', 'V', 'Y', 'U')  /**< bpp=16, Packed mode: Y0+V0+Y1+U0 (1 plane) */
 
+#define SDL_FCC_NV12    SDL_FOURCC('N', 'V', '1', '2')
+
 // RGB formats
 #define SDL_FCC_RV16    SDL_FOURCC('R', 'V', '1', '6')    /**< bpp=16, RGB565 */
 #define SDL_FCC_RV24    SDL_FOURCC('R', 'V', '2', '4')    /**< bpp=24, RGBX8888 */
@@ -63,8 +65,15 @@
 // opaque formats
 #define SDL_FCC__AMC    SDL_FOURCC('_', 'A', 'M', 'C')    /**< Android MediaCodec */
 #define SDL_FCC__VTB    SDL_FOURCC('_', 'V', 'T', 'B')    /**< iOS VideoToolbox */
+#define SDL_FCC__GLES2  SDL_FOURCC('_', 'E', 'S', '2')    /**< let Vout choose format */
 
 // undefine
 #define SDL_FCC_UNDF    SDL_FOURCC('U', 'N', 'D', 'F')    /**< undefined */
+
+enum {
+    IJK_AV_PIX_FMT__START = 10000,
+    IJK_AV_PIX_FMT__ANDROID_MEDIACODEC,
+    IJK_AV_PIX_FMT__VIDEO_TOOLBOX,
+};
 
 #endif